chore(ui): Split `matrix_sdk_ui::room_list::mod.rs` into smaller modules.
No code changes. This patch is just moving things around.
